# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: VideoControl API
  version: 2.5.0
  description: Professional system for centralized media content management
  contact:
    name: VideoControl Team
    url: https://github.com/ya-k0v/VideoControl
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.videocontrol.example.com
    description: Production server

paths:
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - username
                - password
              properties:
                email:
                  type: string
                  format: email
                username:
                  type: string
                password:
                  type: string
                  format: password
                role:
                  type: string
                  enum: [ADMIN, OPERATOR, VIEWER]
      responses:
        '201':
          description: User registered successfully
        '400':
          description: Invalid input
        '409':
          description: User already exists

  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login and get JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Invalid credentials

  /api/devices:
    get:
      tags:
        - Devices
      summary: Get all devices
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of all devices
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  $ref: '#/components/schemas/Device'
    post:
      tags:
        - Devices
      summary: Create new device
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id
                - name
              properties:
                id:
                  type: string
                name:
                  type: string
                type:
                  type: string
      responses:
        '201':
          description: Device created successfully
        '400':
          description: Invalid input
        '409':
          description: Device already exists

  /api/devices/{device_id}:
    get:
      tags:
        - Devices
      summary: Get device by ID
      security:
        - bearerAuth: []
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Device details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '404':
          description: Device not found
    delete:
      tags:
        - Devices
      summary: Delete device
      security:
        - bearerAuth: []
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Device deleted successfully
        '404':
          description: Device not found

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key

  schemas:
    Device:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [DISPLAY, ANDROID_TV, VLC_PLAYER]
        status:
          type: string
          enum: [ONLINE, OFFLINE, ERROR]
        placeholder:
          type: string
          nullable: true
        lastSeen:
          type: string
          format: date-time
        metadata:
          type: object

    File:
      type: object
      properties:
        id:
          type: string
          format: uuid
        deviceId:
          type: string
        fileName:
          type: string
        fileType:
          type: string
          enum: [video, image, pdf, pptx]
        fileSize:
          type: integer
        optimized:
          type: boolean

    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        username:
          type: string
        role:
          type: string
          enum: [ADMIN, OPERATOR, VIEWER]
        isActive:
          type: boolean

    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: object